third_party/blink/web_tests/fast/speechsynthesis/speech-synthesis-speak-multiple-expected.txt

This tests that multiple speak utterances are handled in sequence.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS speechSynthesis.speaking is false
PASS event.elapsedTime >= 0 is true
PASS event.elapsedTime < 1000 is true
PASS speechSynthesis.speaking is true
PASS count is >= i
PASS event.elapsedTime > 0 is true
PASS event.elapsedTime < 10000 is true
PASS speechSynthesis.speaking is true
PASS event.elapsedTime >= 0 is true
PASS event.elapsedTime < 1000 is true
PASS speechSynthesis.speaking is true
PASS count is >= i
PASS event.elapsedTime > 0 is true
PASS event.elapsedTime < 10000 is true
PASS speechSynthesis.speaking is true
PASS event.elapsedTime >= 0 is true
PASS event.elapsedTime < 1000 is true
PASS speechSynthesis.speaking is true
PASS count is >= i
PASS event.elapsedTime > 0 is true
PASS event.elapsedTime < 10000 is true
PASS speechSynthesis.speaking is true
PASS successfullyParsed is true

TEST COMPLETE
